reorg layout, strat, builder, factory
